+++ app/staff/archive/handover/new/page.tsx
@@ -1,4 +1,4 @@
-import ArchiveDocumentFormPage from "@/components/archive/ArchiveDocumentFormPage";
+import ArchiveDocumentFormPage from "@/components/archive/document/ArchiveDocumentFormPage";
 import { archiveDocumentConfigs } from "@/mocks/archiveDocuments";
 
 export default function Page() {
