+++ app/staff/archive/meeting/new/page.tsx
@@ -1,4 +1,4 @@
-import MeetingMinuteFormPage from "@/components/archive/MeetingMinuteFormPage";
+import MeetingMinuteFormPage from "@/components/archive/meeting/MeetingMinuteFormPage";
 
 export default function Page() {
   return <MeetingMinuteFormPage />;
