+++ components/archive/schoolRules/SchoolRulesPage.styles.ts
